Began fixing the AsciiDoc file - split paragraphs
The AsciiDoc conversion was far from perfect.  Its result differs quite
a lot from the original Google Document (PDF).  So started cleaning it
with the goal of making it match the original.  First step was splitting
each paragraph.  All text of a paragraph was on a single line in the
file.  To make it easier to view the forthcoming changes on a paragraph,
each sentence was placed on its own individual line in the file.